Example exercising TLS capabilities with root certificate verification
This example opens an encrypted TLS connection to www.eff.org on port
443. Once the secure connection is negotiated and verified with the
embedded server certificate, this example is very similar to the earlier
raw_async_http_request example for plaintext. The compile-time define
ASYNC_TCP_SSL_ENABLED must be defined and enabled project-wide for this
example to actually enable and use the TLS support.