Prevent deadlock between DNS resolution callback and LWIP call
Stop taking the _asyncsock_mutex when DNS resolution has finished. Doing
so causes a deadlock because the callback runs in the LWIP thread and
therefore will wait forever on the _asyncsock_mutex lock, but the task
holding the lock can in turn attempt a LWIP call, which requires the
LWIP thread to be runnable.