Prevent spurious RX timeout/close on incoming socket object
When accepting a socket from an incoming connection, the internal fd was
assigned without updating the field for last RX activity. If the remote
side did *not* immediately write to the connection, or otherwise the
socket could not be internally checked for incoming data, AND the
higher-level code set up a RX timeout value, the 125-ms interval polling
might evaluate the socket RX timeout condition with a last RX activity
timestamp of zero (as per the default constructor). This caused a
spurious RX timeout and incorrect socket closing. Fixed by initializing
the field when the new incoming connection socket is accepted.